Refactored into a xib
