+++ BracketChallenge/controller/ResultsViewController.swift
@@ -20,11 +20,6 @@ class ResultsViewController: BracketViewController2 {
     override func viewDidDisappear(_ animated: Bool) {
         tabBarController?.navigationItem.rightBarButtonItem = nil
     }
-    
-    //Listeners
-    
-    override func areCellsClickable() -> Bool {
-        //Results are only editable by admins
-        return Identity.user.admin
-    }
+	
+	//TODO: areCellsClickable
 }
